--- electronics_abstract_parts/IdealModel.py
@@ -1,7 +1,0 @@
-from edg_core import *
-
-
-class IdealModel(Block):
-  def contents(self):
-    super().contents()
-    self.require(False, "ideal model")
