+++ electronics_abstract_parts/test_ideal_circuit.py
@@ -2,7 +2,6 @@ import unittest
 
 from electronics_abstract_parts import Resistor
 from electronics_model import *
-from .DummyDevices import DummyVoltageSource
 from .AbstractOpamp import Opamp
 from .OpampCircuits import Amplifier
 
@@ -18,6 +17,12 @@ class AnalogSinkDummy(Block):
   def __init__(self):
     super().__init__()
     self.port = self.Port(AnalogSink(), [InOut])
+
+
+class VoltageDummy(Block):
+  def __init__(self):
+    super().__init__()
+    self.port = self.Port(VoltageSource(), [InOut])
 
 
 class TestOpamp(Opamp):
@@ -44,8 +49,8 @@ class AmplifierTestTop(Block):
     (self.dummyin, ), _ = self.chain(self.dut.input, self.Block(AnalogSourceDummy()))
     (self.dummyref, ), _ = self.chain(self.dut.reference, self.Block(AnalogSourceDummy()))
     (self.dummyout, ), _ = self.chain(self.dut.output, self.Block(AnalogSinkDummy()))
-    (self.dummypwr, ), _ = self.chain(self.dut.pwr, self.Block(DummyVoltageSource()))
-    (self.dummygnd, ), _ = self.chain(self.dut.gnd, self.Block(DummyVoltageSource()))
+    (self.dummypwr, ), _ = self.chain(self.dut.pwr, self.Block(VoltageDummy()))
+    (self.dummygnd, ), _ = self.chain(self.dut.gnd, self.Block(VoltageDummy()))
 
 
 class OpampCircuitTest(unittest.TestCase):
